# openapi.yaml
openapi: 3.1.0
info:
  title: LazyHR API
  description: |
    Human Resources Management System API
    
    **Note**: This API has some path ambiguity between `/api/leave/user/{userId}` and `/api/leave/{leaveId}/*` endpoints.
    In practice, this works because:
    - `user` is a literal string and not interpreted as a leaveId
    - The Spring Boot framework resolves these paths correctly based on the request method and path specificity
    
    However, OpenAPI linting tools may flag these as ambiguous.
  version: 1.0.0
  contact:
    name: LazyHR Team
    email: support@lazyhr.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/lazyhr
    description: Production server
  - url: http://localhost:8080/lazyhr
    description: Development server
...
paths:
...
  /api/leave/apply:
    post:
      tags:
        - Leave Management
      summary: Apply for leave
      description: Submit a new leave request
      operationId: applyForLeave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequestDto'
      responses:
        '200':
          description: Leave application submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/pending:
    get:
      tags:
        - Leave Management
      summary: Get pending leave requests
      description: Retrieve all pending leave requests
      operationId: getPendingLeaveRequests
      responses:
        '200':
          description: Pending leave requests retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch pending requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/stats/pending-count:
    get:
      tags:
        - Leave Management
      summary: Get pending leave requests count
      description: Get the count of pending leave requests
      operationId: getPendingLeaveCount
      responses:
        '200':
          description: Pending count retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch pending count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/user/{userId}:
    get:
      tags:
        - Leave Management
      summary: Get user's leave requests
      description: Retrieve all leave requests for a specific user
      operationId: getUserLeaveRequests
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User leave requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaveRequest'
        '400':
          description: Failed to fetch leave requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/status/{status}:
    get:
      tags:
        - Leave Management
      summary: Get leave requests by status
      description: Retrieve leave requests filtered by status
      operationId: getLeaveRequestsByStatus
      parameters:
        - name: status
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/LeaveStatus'
      responses:
        '200':
          description: Leave requests retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch leave requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/date/{timestamp}:
    get:
      tags:
        - Leave Management
      summary: Get leave requests for a specific date
      description: Retrieve leave requests for a specific date
      operationId: getLeaveRequestsByDate
      parameters:
        - name: timestamp
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Date as Unix timestamp in milliseconds
          example: 1721865600000
      responses:
        '200':
          description: Leave requests for date retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Failed to fetch leave requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/balance/{userId}:
    get:
      tags:
        - Leave Management
      summary: Get leave balance summary
      description: Retrieve leave balance summary for a user
      operationId: getLeaveBalance
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: year
          in: query
          required: false
          schema:
            type: integer
            default: 2024
      responses:
        '200':
          description: Leave balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LeaveBalanceSummary'
        '400':
          description: Failed to fetch leave balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/{leaveId}:
    get:
      tags:
        - Leave Management
      summary: Get leave request by ID
      description: Retrieve a specific leave request by its ID
      operationId: getLeaveRequestById
      parameters:
        - name: leaveId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Leave request retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Leave request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/{leaveId}/approve:
    post:
      tags:
        - Leave Management
      summary: Approve leave request
      description: Approve a pending leave request
      operationId: approveLeaveRequest
      parameters:
        - name: leaveId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: approverId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: comments
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Leave request approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request or leave cannot be approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/{leaveId}/reject:
    post:
      tags:
        - Leave Management
      summary: Reject leave request
      description: Reject a pending leave request
      operationId: rejectLeaveRequest
      parameters:
        - name: leaveId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: approverId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: comments
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Leave request rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request or leave cannot be rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/leave/{leaveId}/cancel:
    delete:
      tags:
        - Leave Management
      summary: Cancel leave request
      description: Cancel a pending leave request
      operationId: cancelLeaveRequest
      parameters:
        - name: leaveId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Leave request cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid request or leave cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # User Management Endpoints - Specific paths first to avoid ambiguity
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: Response status
        message:
          type: string
          description: Response message
        data:
          type: object
          description: Response data
      required:
        - status
        - message
    LeaveRequestDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: ID of the user requesting leave
        leaveCategory:
          $ref: '#/components/schemas/LeaveCategory'
        leavePeriod:
          $ref: '#/components/schemas/LeavePeriod'
        startDate:
          type: integer
          format: int64
          description: Start date of leave (Unix timestamp in milliseconds)
          example: 1721865600000
        endDate:
          type: integer
          format: int64
          description: End date of leave (Unix timestamp in milliseconds)
          example: 1722038400000
        reason:
          type: string
          description: Reason for leave request
        totalDays:
          type: number
          format: double
          description: Total number of days requested
      required:
        - userId
        - leaveCategory
        - leavePeriod
        - startDate
        - endDate
        - reason
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        leaveCategory:
          $ref: '#/components/schemas/LeaveCategory'
        leavePeriod:
          $ref: '#/components/schemas/LeavePeriod'
        startDate:
          type: integer
          format: int64
          description: Start date of leave (Unix timestamp in milliseconds)
          example: 1721865600000
        endDate:
          type: integer
          format: int64
          description: End date of leave (Unix timestamp in milliseconds)
          example: 1722038400000
        reason:
          type: string
        status:
          $ref: '#/components/schemas/LeaveStatus'
        appliedDate:
          type: integer
          format: int64
          description: When the leave was applied (Unix timestamp in milliseconds)
          example: 1721779200000
        totalDays:
          type: number
          format: double
        approvedBy:
          $ref: '#/components/schemas/User'
        approvedDate:
          type: integer
          format: int64
          description: When the leave was approved/rejected (Unix timestamp in milliseconds)
          example: 1721865600000
          nullable: true
        comments:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        hireDate:
          type: integer
          format: int64
          description: Date of hire (Unix timestamp in milliseconds)
          example: 1609459200000
          nullable: true
        salary:
          type: number
          format: double
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
        createdAt:
          type: integer
          format: int64
          description: User creation timestamp (Unix timestamp in milliseconds)
          example: 1709251200000
        updatedAt:
          type: integer
          format: int64
          description: User last update timestamp (Unix timestamp in milliseconds)
          example: 1721779200000
...
    LeaveBalanceSummary:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        year:
          type: integer
        annualLeaveAllocated:
          type: integer
        annualLeaveUsed:
          type: integer
        sickLeaveAllocated:
          type: integer
        sickLeaveUsed:
          type: integer
        privateLeaveAllocated:
          type: integer
        privateLeaveUsed:
          type: integer
    LeaveCategory:
      type: string
      enum:
        - ANNUAL
        - SICK
        - SPECIAL_HOLIDAY
        - PRIVATE
      description: Type of leave being requested
    LeavePeriod:
      type: string
      enum:
        - FULL_DAY
        - AM
        - PM
      description: Period of leave (full day, morning, or afternoon)
    LeaveStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - CANCELLED
      description: Current status of the leave request
    Role:
      type: string
      enum:
        - ADMIN
        - MANAGER
        - EMPLOYEE
      description: User role in the system
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []